Firebase Config update
Apparenlty the firebase config description has to be put into both app.component and app.module for the auth system and the feed system to work. I was trying to imort the config from the component into the module, but I could not get it and Ahmed is no longer here to help explain his code.